# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Issue Tracker - API
  version: 1.0.0
  description: OpenAPI-like summary for the local AI Issue Tracker API routes
servers:
  - url: http://localhost:3000
paths:
...
  /api/ai/generate-resolution:
    post:
      summary: Generate AI resolution (RAG + Ollama)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                issueId:
                  type: string
      responses:
        '200':
          description: Created resolution
...